Refactor usechatflow
